+++ src/lr-wpan/model/lr-wpan-fields.cc
@@ -492,6 +492,39 @@ CapabilityField::CapabilityField()
     m_allocAddr = true;
 }
 
+CapabilityField::CapabilityField(uint8_t bitmap)
+{
+    SetCapability(bitmap);
+}
+
+uint8_t
+CapabilityField::GetCapability() const
+{
+    uint8_t capability;
+
+    capability = (m_reservedBit0) & (0x01);                  //!< Bit 0 (reserved)
+    capability |= (m_deviceType << 1) & (0x01 << 1);         //!< Bit 1
+    capability |= (m_powerSource << 2) & (0x01 << 2);        //!< Bit 2
+    capability |= (m_receiverOnWhenIdle << 3) & (0x01 << 3); //!< Bit 3
+    capability |= (m_reservedBit45 << 4) & (0x03 << 4);      //!< Bit 4-5 (reserved)
+    capability |= (m_securityCap << 6) & (0x01 << 6);        //!< Bit 6
+    capability |= (m_allocAddr << 7) & (0x01 << 7);          //!< Bit 7
+
+    return capability;
+}
+
+void
+CapabilityField::SetCapability(uint8_t bitmap)
+{
+    m_reservedBit0 = (bitmap) & (0x01);            //!< Bit 0 (reserved)
+    m_deviceType = (bitmap >> 1) & (0x01);         //!< Bit 1
+    m_powerSource = (bitmap >> 2) & (0x01);        //!< Bit 2
+    m_receiverOnWhenIdle = (bitmap >> 3) & (0x01); //!< Bit 3
+    m_reservedBit45 = (bitmap >> 4) & (0x03);      //!< Bit 4-5 (reserved)
+    m_securityCap = (bitmap >> 6) & (0x01);        //!< Bit 6
+    m_allocAddr = (bitmap >> 7) & (0x01);          //!< Bit 7
+}
+
 uint32_t
 CapabilityField::GetSerializedSize() const
 {
@@ -501,16 +534,7 @@ CapabilityField::GetSerializedSize() const
 Buffer::Iterator
 CapabilityField::Serialize(Buffer::Iterator i) const
 {
-    uint8_t capability;
-
-    capability = 0;                                          //!< Bit 0 (reserved)
-    capability = (m_deviceType << 1) & (0x01 << 1);          //!< Bit 1
-    capability |= (m_powerSource << 2) & (0x01 << 2);        //!< Bit 2
-    capability |= (m_receiverOnWhenIdle << 3) & (0x01 << 3); //!< Bit 3
-                                                             //!< Bit 4-5 (reserved)
-    capability |= (m_securityCap << 6) & (0x01 << 6);        //!< Bit 6
-    capability |= (m_allocAddr << 7) & (0x01 << 7);          //!< Bit 7
-    i.WriteU8(capability);
+    i.WriteU8(GetCapability());
     return i;
 }
 
@@ -518,14 +542,7 @@ Buffer::Iterator
 CapabilityField::Deserialize(Buffer::Iterator i)
 {
     uint8_t capability = i.ReadU8();
-    //!< Bit 0 (reserved)
-    m_deviceType = (capability >> 1) & (0x01);         //!< Bit 1
-    m_powerSource = (capability >> 2) & (0x01);        //!< Bit 2
-    m_receiverOnWhenIdle = (capability >> 3) & (0x01); //!< Bit 3
-                                                       //!< Bit 4-5 (reserved)
-    m_securityCap = (capability >> 6) & (0x01);        //!< Bit 6
-    m_allocAddr = (capability >> 7) & (0x01);          //!< Bit 7
-
+    SetCapability(capability);
     return i;
 }